Create separate AddressTable and CardTable components from ProfilePage
