+++ frontend/src/components/AddressTable.less
@@ -1,0 +1,7 @@
+.bookstore-address-table {
+  white-space: nowrap;
+}
+
+.bookstore-address-table-expanded-container {
+  white-space: normal;
+}
