+++ frontend/src/components/CardTable.less
@@ -1,0 +1,13 @@
+.bookstore-credit-card-table {
+  white-space: nowrap;
+}
+
+.bookstore-credit-card-table-expanded-container {
+  display: flex;
+  justify-content: space-between;
+  white-space: normal;
+}
+
+.bookstore-credit-card-table-expanded-container > div:last-of-type {
+  margin: 0px 32px;
+}
