+++ frontend/src/pages/ProfilePage.less
@@ -47,12 +47,3 @@
 .rccs {
   margin: 0px !important;
 }
-
-.bookstore-credit-card-table-expanded-container {
-  display: flex;
-  justify-content: space-between;
-}
-
-.bookstore-credit-card-table-expanded-container > div:last-of-type {
-  margin-right: 64px;
-}
